refactor: move AssetID class to _asset_id module and update imports
- Relocated AssetID class from _entity.py to _asset_id.py for better module organization.
- Updated import statements in test file to reflect the new location of AssetID.
- Adjusted module docstring in the test file to match the new structure.
